backend/requirements.txt
# backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0

# Authentication
djangorestframework-simplejwt==5.3.0

# CORS handling
django-cors-headers==4.3.1

# Filtering and pagination
django-filter==23.3

# Development tools
django-debug-toolbar==4.2.0

# Production server (optional)
gunicorn==21.2.0

# Environment variables
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.10.1